% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_box_scores}
\alias{get_box_scores}
\title{Box Scores Scrape}
\usage{
get_box_scores(
  game_ids,
  multi.games = F,
  use_file = F,
  save_file = F,
  base_path = NA,
  overwrite = F
)
}
\arguments{
\item{multi.games}{whether to aggregate over games}

\item{game_id}{vector of box ids for the game}
}
\value{
data frame with each row representing a player in the game
}
\description{
This function returns a single data frame for a vector of box ids. A wrapping of scrape_box for multiple games
}
\examples{
get_box_scores(c(1982642, 1982641))
}
