% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_date_games}
\alias{get_date_games}
\title{Date Schedule Scrape}
\usage{
get_date_games(
  date = as.character(format(Sys.Date() - 1, "\%m/\%d/\%Y")),
  conference = "All",
  conference.ID = NA,
  use_file = F,
  save_file = F,
  base_path = NA
)
}
\arguments{
\item{date}{a character object containing a date in the format mm/dd/yyyy. Defaults is previous day (yesterday)}

\item{conference}{the common name used for a conference, not sensitive to case, spacing, punctuation, etc.}

\item{conference.ID}{alternatively, if the conference ID is known it replace the conference name variable}
}
\value{
data frame with each row representing an inidividual game
\itemize{
\item{Date} - Game date
\item{Start_Time} - Start time reported in eastern time zone (I believe)
\item{Home} - Home team
\item{Away} - Away team
\item{GameID} - If the game is finished and has play-by-play data available, the game ID used to scrape game data
\item{Home_Score} - If the game is finished, the final score for the home team
\item{Away_Score} - If the game is finished, the final score for the away team
\item{Attendance} - The attendance count reported by the NCAA
\item{Neutral_Site} - A logical variable that is true when the game was played at a designated neutral destination
}
}
\description{
This function returns a schedule for the given date and specified conference.
Results are included if applicable, as well as the play-by-play game id
}
\examples{
get_date_games(date = "12/11/2018", conference = "Ivy")
get_date_games(date = "12/11/2018", conference.ID = 865)
}
