% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_play_by_play}
\alias{get_play_by_play}
\title{Multiple Game Play-By-Play Scraper}
\usage{
get_play_by_play(game_ids)
}
\arguments{
\item{game_ids}{A string/numeric object or vector containing game ids}
}
\description{
This is the suggested function for scraping play-by-play, as it handles and binds multiple games.
Please see the scrape_game() function documentation for a more detailed description.
}
\examples{
get_play_by_play(c(4671170, 4674164))
team_sched <- get_team_schedule(team.id = 450680)
get_play_by_play(team_sched$Game_ID)
}
