% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_player_lineups}
\alias{get_player_lineups}
\title{Player Lineup Finder}
\usage{
get_player_lineups(Lineup_Data = NA, Included = NA, Excluded = NA)
}
\arguments{
\item{Lineup_Data}{a data frame of lineups created from get_lineups()}

\item{Included}{a character vector of players to be included in all lineups}

\item{Excluded}{a character vector of players to be excluded from all lineups}
}
\value{
data frame of lineups with statistics documented in get_lineups()
}
\description{
This function finds all lineups from a given lineup data source that include/exclude certain players. It acts
as a quick way to filter lineups for players
}
