% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_player_stints.R
\name{get_player_stints}
\alias{get_player_stints}
\title{Helper function to get player stints
Used in plot_player_stints}
\usage{
get_player_stints(game_data, player_name)
}
\description{
Helper function to get player stints
Used in plot_player_stints
}
