% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_possessions}
\alias{get_possessions}
\title{Possession Parsing}
\usage{
get_possessions(play_by_play_data = NA, simple = F)
}
\arguments{
\item{play_by_play_data}{dataframe of play by play data from get_play_by_play}
}
\value{
data frame with each row representing a possession
}
\description{
This function aggregates play by play data into individual possession results
}
