% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_team_roster}
\alias{get_team_roster}
\title{Team Roster Scrape}
\usage{
get_team_roster(team.id = NA, season = NA, team.name = NA)
}
\arguments{
\item{team.id}{The unique id given to each college/team for each season. This can be found in the url of the team page.}

\item{season}{Alternative to using the id, you can get a team from data(ids) with a season and team name specification.
String for the season stored as yyy1-y2 (2018-19 is current)}

\item{team.name}{Alternative to using the id, you can get a team from data(ids) with a season and team name specification.
This inputs a team name, to be used along with season. This needs the school name not the complete team name, so "Duke" not "Duke Blue Devils".}
}
\value{
data frame with each row representing a player on the roster
\itemize{
\item{Jersey} - Player jersey number
\item{Player} - Name of Player
\item{Pos} - Position (one of G,F,C) as designated by the NCAA
\item{Ht} - Height as reported by the NCAA
\item{Yr} - School year, as Fr, So, Jr, Sr
}
}
\description{
This function returns a data frame of the roster for the specified team. This will include player names and positions
as well as jersey number, height and school year. Note: currently,
the season/team.name parameters can only be used for the 2016-17, 2017-18, 2018-19 seasons
}
\examples{
get_team_roster(team.id = 450680)
get_team_roster(season = "2018-19", team.name = "Penn")
}
