% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_team_schedule}
\alias{get_team_schedule}
\title{Team Schedule Scrape}
\usage{
get_team_schedule(team.id = NA, season = NA, team.name = NA)
}
\arguments{
\item{team.id}{The unique id given to each college/team for each season. This can be found in the url of the team page.}

\item{season}{Season following format yyy1-y2, ex "2018=19"}

\item{team.name}{Alternative to using the id, you can get a team from data(ids) with a season and team name specification.
This inputs a team name, to be used along with season. This needs the school name not the complete team name, so "Duke" not "Duke Blue Devils".}
}
\value{
data frame with each row representing an individual game
\itemize{
\item{Date} - Game date
\item{Home} - Home team
\item{Home_Score} - If the game is finished, the final score for the home team
\item{Away} - Away team
\item{Away_Score} - If the game is finished, the final score for the away team
\item{Game_ID} - If the game is finished and has play-by-play data available, the game ID used to scrape game data
\item{isNeutral} - A logical variable that is true when the game was played at a designated neutral destination
\item{Detail} - Additional detail such as if the game went into OT and # of OTs
}
}
\description{
This function returns a data frame of the schedule for the specified team. This will include game ids used
for play-by-play scraping if the game has ended, along with the team scores and attendance. Note: currently,
the season/team.name parameters can only be used for the 2016-17, 2017-18, 2018-19 seasons
}
\examples{
get_team_schedule(team.id = 450680)
get_team_schedule(season = "2018-19", team.name = "Penn")
}
