% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_team_stats}
\alias{get_team_stats}
\title{Team Stats Function}
\usage{
get_team_stats(play_by_play_data = NA, include_transition = F)
}
\arguments{
\item{play_by_play_data}{data frame consisting of play-by-play data from the functions scrape_game() or get_play_by_play()}
}
\value{
data frame with each row representing a team
}
\description{
This function takes in pbp data and calculates team level stats
}
