% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{plot_duos}
\alias{plot_duos}
\title{Duo Plot}
\usage{
plot_duos(Lineup_Data = NA, team = NA, min_mins = 0, regressed_poss = 50)
}
\arguments{
\item{team}{a single team name}

\item{min_mins}{minimum number of minutes required for duo to play together}

\item{regressed_poss}{effectively number of team average possessions used to "shrink" ratings}

\item{lineup_data}{a data frame of lineups created from get_lineups()}
}
\value{
ggplot object of network plot
}
\description{
Creates network plot of team ratings with players together on the court
Uses empirical bayesian formula to estimate ratings relative to team average:
regressed ORTG = (n_bar*r_bar_o + PTS) / (n_bar + POSS)
n_bar = average possessions per lineup + regressed possessions
r_bar_o = team offensive rating (PTS/POSS)*100
}
