% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{plot_mins_dist}
\alias{plot_mins_dist}
\title{Minutes Distribution Plot}
\usage{
plot_mins_dist(play_by_play_data = NA, team = NA, threshold = NA)
}
\arguments{
\item{team}{a single team name}

\item{threshold}{minimum minutes played to include in plot}

\item{pbp_data}{a data frame of pbp created from get_play_by_play() / scrape_game()}
}
\value{
ggplot object containing minutes distribution
}
\description{
This function takes in play by play data and a team name and returns a plot
showing the distributon of when each player was on the court
}
