% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_player_stints.R
\name{plot_player_stints}
\alias{plot_player_stints}
\title{Stint Plot}
\usage{
plot_player_stints(play_by_play_data = NA)
}
\arguments{
\item{play_by_play_data}{a data frame of lineups created from get_lineups(). Must be a single game}
}
\value{
ggplot object of stint plot
}
\description{
Plots all player and lineup stints for a single game, along with player ending +/-
Plot is colored by net points added/lost during stint
}
