% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{scrape_game}
\alias{scrape_game}
\title{Individual Game Play-By-Play Scraper}
\usage{
scrape_game(game_id)
}
\arguments{
\item{game_id}{string made up of digits given to each unique game. This can be found in the play-by-play url for each game.}
}
\value{
data frame containing play-by-play data for a game, where each row represents an individual event from the game.
\itemize{
\item{ID} - Numeric game id that is given for each unique game
\item{Date} - Game date
\item{Home} - Home team name
\item{Away} - Away team name
\item{Time} - String for game time in format reported originally by NCAA
\item{Game_Time} - String reporting game time elapsed. For example, a normal game starts at 00:00 and ends at 40:00
\item{Game_Seconds} - Number of seconds elapsed in the game
\item{Half_Status} - Number indicating which half: 1,2 for regulation, 3+ for OTs
\item{Home_Score} - Score for home team after event occurred as reported by NCAA
\item{Away_Score} - Score for away team after event occurred as reported by NCAA
\item{Event_Team} - Which team was repsonsibile for the play-by-play entry
\item{Event_Description} - The text description of the event on the NCAA site
\item{Player_1} - The primary player responsible for the event
\item{Player_2} - The secondary player within an event. As of now this is only the assister on a made shot.
\item{Event_Type} - String representing the event that occurred as reported by the description
\item{Event_Result} - Takes values of made/missed for shot attempts, otherwise NA
\item{Shot_Value} - Numeric value of points awarded by shot type. Ranges from 1-3 for shot attempts, otherwise NA
\item{Event_Length} - Estimate of time before events calculated from event time - last previous event time
\item{Home.1} - One of the players on the court for the home team
\item{Home.2} - One of the players on the court for the home team
\item{Home.3} - One of the players on the court for the home team
\item{Home.4} - One of the players on the court for the home team
\item{Home.5} - One of the players on the court for the home team
\item{Away.1} - One of the players on the court for the away team
\item{Away.2} - One of the players on the court for the away team
\item{Away.3} - One of the players on the court for the away team
\item{Away.4} - One of the players on the court for the away team
\item{Away.5} - One of the players on the court for the away team
\item{Status} - This reports the cleanliness of the data frame. Will be CLEAN if no errors are found. Otherwise will say
the number of errors that occurred or a potential substitution mistake occurred.
}
}
\description{
This function retrieves and cleans play-by-play information for an individual game.
Warns users of potential errors and mistakes made by the game trackers. The number of player discrepancies warning
counts displays the number of events players committed when it is found they were not on the court at the time of the
event. The substitution mistake warning indicates an unclean substitution was entered. (ex. 2 players enter and 1 leaves)
}
\examples{
scrape_game(4674164)
}
