% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{get_lineups}
\alias{get_lineups}
\title{Lineup Compiler}
\usage{
get_lineups(
  play_by_play_data = NA,
  keep.dirty = F,
  garbage.filter = F,
  error.thresh = 5
)
}
\arguments{
\item{play_by_play_data}{data frame consisting of play-by-play data from the functions scrape_game() or get_play_by_play()}

\item{keep.dirty}{logical variable to specify whether or not to filter out potentially inaccurrate data. When FALSE,
will remove all rows from games where the number of discrepencies is above the desired count. Defaults to FALSE.}

\item{garbage.filter}{logical variable to specify whether or not to filter out garbage time entries. Uses format similar
to cleaningtheglass.com definitions, but adjusted because of different time/scoring format. This is very subjective with no
true criteria or reasoning. Follows rule: a game that enters garbage time cannot exit even if a team comes back.
The remainder of a game is garbage time once a team is: Up by 25+ with 5 mins gone in 2nd half, up 20 with 10 mins gone in 2nd half,
up 15 with 16 mins gone in 2nd half. Defaults to FALSE, which means garbage time will not be removed}

\item{error.thresh}{numeric variable that lets user set their preferred discrepancy threshold with the keep.dirty variable. This means
when less than the threshold occurs in a game, it will be considered clean. As defined in scrape_game(), a discrepancy occurs when a
player registers an event when they are not found to be on the court. Defaults to 5 discrepancies.}
}
\value{
data frame with each row representing a unique lineup. All stats for entire lineup or opponent (o- prefix)
\itemize{
\item{P1} - Player in lineup
\item{P2} - Player in lineup
\item{P3} - Player in lineup
\item{P4} - Player in lineup
\item{P5} - Player in lineup
\item{Team} - Team for the lineup
\item{Mins} - Minutes the lineup was on the court
\item{PTS} - Points scored
\item{FGA} - Field goal attempts
\item{TO} - Turnovers
\item{TPA} - Three point attempts
\item{FGM} - Field goals made
\item{TPM} - Three points made
\item{FTA} - Free throw attempts
\item{FTM} - Free throws made
\item{ORB} - Offensive rebounds
\item{DRB} - Defensive rebounds
\item{RIMA} - Rim attempts: defined as layups, dunks, tip-ins, hook attempts
\item{BLK} - Blocked shots
\item{AST} - Assists
\item{POSS} - (Offensive) Possessions: (FGA + .475 x FTA - ORB + TO + oFGA + .475 x oFTA - oORB + oTO) /2
\item{TS.} - True shooting percentage: (PTS / 2) / (FGA + .475 x FTA),
\item{eFG.} - Effective field goal percentage: (FGM + 0.5 x TPM) / FGA,
\item{TPP} - Three point percentage: TPA/TPM
\item{FTP} - Free throw percentage: FTA/FTM
\item{TPrate} - Three point attempt rate: TPA/FGA
\item{ASTrate} - Assist rate: AST/FGM
\item{TOrate} - Turnover rate: TO/POSS
\item{FTrate} - Free throw rate: FTA/FGA
\item{BLKrate} - Block rate: BLK/FGA
\item{ORB.} - Offensive rebound percentage: ORB / (ORB + oDRB)
\item{DRB.} - Defensive rebound percentage: DRB / (DRB + oORB)
\item{OEFF} - Offensive efficiency: 100 * (PTS/POSS)
\item{DEFF} - Defensive efficiency: 100 * (oPTS/POSS)
\item{NETEFF} - Net efficiency: OEFF - DEFF
\item{PACE} - Average time per possession (Seconds): (Possessions / Mins) * 60
\item{ShotsPerPoss} - Estimate of shot attempts per possession: 1 + (ORB - TO) / POSS,
}
}
\description{
This function takes in a play-by-play dataframe, and generates all possible lineups for both teams.
It then calculates a variety of statistics/metrics at a lineup level.
}
