% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_box_score.R
\name{scrape_box_score}
\alias{scrape_box_score}
\title{Get Box Score Data per game}
\usage{
scrape_box_score(game_id)
}
\arguments{
\item{game_id}{game_id

Game ID can be retrived from the schedule function}
}
\description{
Get Box Score Data per game
}
\examples{


}
\keyword{internal}
